requirements.txt
albumentations==1.3.1
fvcore==0.1.5.post20221221
matplotlib==3.9.2
numpy==1.26.4
timm==1.0.11
Pillow
